+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [2026.06.19.1357] - 2026-06-19
+
+- Automated build for commit e1160f6f61660cd329724dc0a41cdc6c83cf03f1.
+
+## [2026.06.19.1354] - 2026-06-19
+
+- Automated build for commit e1160f6f61660cd329724dc0a41cdc6c83cf03f1.
+
+## [2026.06.19.1353] - 2026-06-19
+
+- Automated build for commit e1160f6f61660cd329724dc0a41cdc6c83cf03f1.
+
+## [2026.06.19.1350] - 2026-06-19
+
+- Automated build for commit e1160f6f61660cd329724dc0a41cdc6c83cf03f1.
+
 ## [2026.06.19.1324] - 2026-06-19
 
 - Automated build for commit cd8febc6e303658fbde4dc605cc218d15c2ec0df.